Move history context definition to useSyncHistory
